requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
decorator==4.3.0
defusedxml==0.5.0
Django==2.1
django-appconf==1.0.2
django-autocomplete-light==3.3.0
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-widget-tweaks==1.4.2
eight==0.4.2
future==0.16.0
idna==2.7
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
lxml==4.2.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
pyOpenSSL==18.0.0
python-stdnum==1.9
pytz==2018.5
reportlab==3.5.6
requests==2.19.1
signxml==2.5.2
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
